Fixed broken specs, renamed infrastructure directories and added extra validation inside user id
