+++ spec/BenGorUser/User/Application/Service/SignUp/SignUpUserHandlerSpec.php
@@ -24,7 +24,7 @@ use BenGorUser\User\Domain\Model\UserId;
 use BenGorUser\User\Domain\Model\UserPassword;
 use BenGorUser\User\Domain\Model\UserRepository;
 use BenGorUser\User\Domain\Model\UserRole;
-use BenGorUser\User\Infrastructure\Security\Test\DummyUserPasswordEncoder;
+use BenGorUser\User\Infrastructure\Security\DummyUserPasswordEncoder;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
