+++ src/BenGorUser/User/Domain/Model/Exception/UserIdInvalidException.php
@@ -1,0 +1,23 @@
+<?php
+
+/*
+ * This file is part of the BenGorUser package.
+ *
+ * (c) Beñat Espiña <benatespina@gmail.com>
+ * (c) Gorka Laucirica <gorka.lauzirika@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace BenGorUser\User\Domain\Model\Exception;
+
+/**
+ * User id invalid exception class.
+ *
+ * @author Beñat Espiña <benatespina@gmail.com>
+ * @author Gorka Laucirica <gorka.lauzirika@gmail.com>
+ */
+final class UserIdInvalidException extends \InvalidArgumentException
+{
+}
